+++ VersionC/PCElements/GICLine.cpp
@@ -620,7 +620,7 @@ void TGICLineObj::GetVterminalForSource()
   // non-harmonic modes or no spectrum
 			{
 				int stop = 0;
-				if(Abs((int) (with0->get_FFrequency() - SrcFrequency)) > EPSILON2)
+				if(Abs((with0->get_FFrequency() - SrcFrequency)) > EPSILON2)
 					Vmag = 0.0;  // Solution Frequency and Source Frequency don't match!
 					       /*NOTE: RE-uses VTerminal space*/
 				for(stop = Fnphases, i = 1; i <= stop; i++)
